perf(pipelineRef): define deserializer as CustomTriggerDeserializerSupplier and move changes in orca-core to orca-sql.
